added simple scrolling and highlights
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@2371 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ menu/scroll_list.h
@@ -25,7 +25,7 @@ private:
 	List _list;
 
 	int _pos;
-	size_t _current_item;
+	int _current_item;
 };
 
 #endif
